+++ +gams/+transfer/+symbol/Equation.m
@@ -202,7 +202,7 @@ classdef Equation < gams.transfer.symbol.Symbol
                 error(e.message);
             end
             if ~has_records
-                obj.data_ = gams.transfer.symbol.data.Struct();
+                obj.data_ = gams.transfer.symbol.data.Struct.Empty(obj.def_.domains);
             end
         end