+++ +gams/+transfer/+symbol/Parameter.m
@@ -157,7 +157,7 @@ classdef Parameter < gams.transfer.symbol.Symbol
                 error(e.message);
             end
             if ~has_records
-                obj.data_ = gams.transfer.symbol.data.Struct();
+                obj.data_ = gams.transfer.symbol.data.Struct.Empty(obj.def_.domains);
             end
         end